requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.4.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0